a program that finds and prints the largest prime factor of the number 612852475143

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,8 +8,8 @@
 
 int main(void)
 {
-	unsigned long long int number = 612852475143;
-	unsigned long long int prime_factor;
+	unsigned long int number = 612852475143;
+	unsigned long int prime_factor;
 
 	for (prime_factor = 2; prime_factor <= number; prime_factor++)
 	{
